update stepdefs folder
